script upgrades runallstates.py runs nicely multithreaded. adds --dry-run to just print commands. runallstates.pl extra-args, but will be deprecated soon. manybest.pl also parses older results. bahtml.pl copies statsum into html. measureRatio.pl uses LL/ll.gbin instead of just LL/.gbin

+++ runallstates.pl
@@ -8,10 +8,11 @@ if ( ! defined $datadir ) {
 }
 $bindir = undef;
 $exe = undef;
-$stdargs = " --blankDists --sLog g/ --statLog statlog --d2 ";
+$stdargs = " --blankDists --sLog g/ --statLog statlog ";
 $start = time();
 $end = undef;
 @statearglist = ();
+$extrargs = "";
 
 while ( $arg = shift ) {
 	if ( $arg eq "--data" ) {
@@ -33,6 +34,8 @@ while ( $arg = shift ) {
 	} elsif ( $arg eq "--runsecs" ) {
 		$arg = shift;
                 $end = $start + $arg;
+	} elsif ( $arg eq "--" ) {
+		$extrargs = " " . shift;
 	} else {
 		local $astu = uc($arg);
 		if ( -r "$datadir/$astu/basicargs" ) {
@@ -136,51 +139,57 @@ sub shouldstop() {
 
 while ( ! shouldstop() ) {
 	sleep 1;
-        ILP: foreach $stu ( @states ) {
+	ILP: foreach $stu ( @states ) {
 	if ( ! -d "${stu}" ) {
 		mkdir "${stu}" or die "could not mkidr \"${stu}\": !$\n";
 	}
-        $ha = "";
-        # re-read handargs every time in case they change.
-        $haname = "$datadir/$stu/handargs";
-        if ( -r $haname ) {
-          open(FIN, '<', $haname) or die "$haname: $!";
-          $ha = <FIN>;
-          close FIN;
-          chomp $ha;
-          $ha = " " . $ha;
-        }
+	$ha = "";
+	# re-read handargs every time in case they change.
+	$haname = "$datadir/$stu/handargs";
+	if ( -r $haname ) {
+		open(FIN, '<', $haname) or die "$haname: $!";
+		$ha = <FIN>;
+		close FIN;
+		chomp $ha;
+		$ha = " " . $ha;
+	}
 	$ctd = $stu . "/" . timestamp();
 	mkdir( $ctd ) or die "could not mkdir $ctd: $!\n";
 	mkdir( "$ctd/g" ) or die "could not mkdir $ctd/g: $!\n";
 	chdir( $ctd ) or die "could not chdir to $ctd: $!\n";
-        $cmd = $nice . $exe . $stdargs . $basicargs{$stu} . $ha;
-        print $cmd . "\n";
+	# Need to create a dummy regular file because of how kerberized perms work
+	open( FOUTT, '>', "statsum" ) or die "could not create statsum: $!\n";
+	close FOUTT;
+	$cmd = $nice . $exe . $stdargs . $basicargs{$stu} . $ha . $extrargs;
+	print $cmd . "\n";
 	system $cmd;
+	if ( ! -f "statlog" ) {
+		die "run left no statlog. crazy.\n";
+	}
 	system "grep ^# statlog > statsum";
 	system "gzip statlog";
 # tar saves a lot of space vs the many small files collected and removed
 	system "tar jcf g.tar.bz2 g";
 	system "rm -rf g";
-        if ( (defined $bindir) && (-x "$bindir/manybest.pl") && chdir("..") ) {
-          system( "$bindir/manybest.pl", "-ngood", "15", "-mvbad", "-rmbad", "-rmempty", "-n", "10" );
-	  $drendcmd = $drendcmds{$stu};
-	  if ( (defined $drendcmd) && (! -f "link1/${stu}_final2.png") ) {
-	    print $drendcmd . "\n";
-	    system $drendcmd;
-	    $cmd = "convert $datadir/$stu/${stu}_start.png link1/${stu}_final2.png +append link1/${stu}_ba.png";
-	    print $cmd . "\n";
-	    system $cmd;
-	    $cmd = "convert link1/${stu}_ba.png -resize 500x500 link1/${stu}_ba_500.png";
-	    print $cmd . "\n";
-	    system $cmd;
-	  }
-        }
+	if ( (defined $bindir) && (-x "$bindir/manybest.pl") && chdir("..") ) {
+		system( "$bindir/manybest.pl", "-ngood", "15", "-mvbad", "-rmbad", "-rmempty", "-n", "10" );
+		$drendcmd = $drendcmds{$stu};
+		if ( (defined $drendcmd) && (! -f "link1/${stu}_final2.png") ) {
+			print $drendcmd . "\n";
+			system $drendcmd;
+			$cmd = "convert $datadir/$stu/${stu}_start.png link1/${stu}_final2.png +append link1/${stu}_ba.png";
+			print $cmd . "\n";
+			system $cmd;
+			$cmd = "convert link1/${stu}_ba.png -resize 500x500 link1/${stu}_ba_500.png";
+			print $cmd . "\n";
+			system $cmd;
+		}
+	}
 # TODO: $datadir/$stu/drendcmd line 1, modified for datadir, bindir and link1
 	chdir( $rootdir ) or die "could not chdir to $rootdir: $!\n";
-        if ( shouldstop() ) {
-          last ILP;
-        }
-        }
+	if ( shouldstop() ) {
+		last ILP;
+	}
+	}
 }
 unlink "stop";
